# Documentation/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal and family finance management system with multi-currency support and automatic balance calculation",
        "title": "Expense Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Igor Kudinov",
            "email": "digitlock@proton.me"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/reports/monthly-summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Monthly financial summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month YYYY-MM (defaults to current)",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.MonthlySummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/reports/spending-by-category": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Spending by category report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date YYYY-MM-DD",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date YYYY-MM-DD",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "expense",
                        "description": "income or expense",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SpendingByCategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AccountBalances": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    },
                    "example": {
                        "Bank Card RSD": 206700,
                        "Cash Wallet RSD": 2000
                    }
                },
                "total": {
                    "type": "number",
                    "example": 208700
                }
            }
        },
...
        "dto.CategorySpending": {
            "type": "object",
            "properties": {
                "average_per_transaction": {
                    "type": "number",
                    "example": 1437.5
                },
                "category_id": {
                    "type": "string",
                    "example": "e3b5cae7-cf1d-4c22-a0fe-4014590a7a85"
                },
                "category_name": {
                    "type": "string",
                    "example": "Groceries"
                },
                "percentage": {
                    "type": "number",
                    "example": 19.5
                },
                "total_amount": {
                    "type": "number",
                    "example": 11500
                },
                "transaction_count": {
                    "type": "integer",
                    "example": 8
                }
            }
        },
...
        "dto.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ValidationError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorDetail"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.MonthlySummary": {
            "type": "object",
            "properties": {
                "net_savings": {
                    "type": "number",
                    "example": 16025
                },
                "savings_rate": {
                    "type": "number",
                    "example": 21.4
                },
                "total_expenses": {
                    "type": "number",
                    "example": 58975
                },
                "total_income": {
                    "type": "number",
                    "example": 75000
                }
            }
        },
        "dto.MonthlySummaryResponse": {
            "type": "object",
            "properties": {
                "account_balances": {
                    "$ref": "#/definitions/dto.AccountBalances"
                },
                "currency": {
                    "type": "string",
                    "example": "RSD"
                },
                "currency_note": {
                    "description": "CurrencyNote is set only when a non-RSD currency was requested but no\nexchange rate was available, so amounts fall back to RSD.",
                    "type": "string",
                    "example": "Requested EUR, but no exchange rate available; showing RSD"
                },
                "expense_breakdown": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    },
                    "example": {
                        "Groceries": 11500,
                        "Utilities": 8500
                    }
                },
                "generated_at": {
                    "type": "string",
                    "example": "2025-12-06T18:30:00Z"
                },
                "income_breakdown": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    },
                    "example": {
                        "Freelance": 25000,
                        "Salary": 50000
                    }
                },
                "month": {
                    "type": "string",
                    "example": "2025-11"
                },
                "report_type": {
                    "type": "string",
                    "example": "monthly_summary"
                },
                "summary": {
                    "$ref": "#/definitions/dto.MonthlySummary"
                },
                "transaction_counts": {
                    "$ref": "#/definitions/dto.TransactionCounts"
                }
            }
        },
...
        "dto.ReportPeriod": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2025-11-30"
                },
                "start_date": {
                    "type": "string",
                    "example": "2025-11-01"
                }
            }
        },
        "dto.SpendingByCategoryResponse": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string",
                    "example": "RSD"
                },
                "currency_note": {
                    "description": "CurrencyNote is set only when a non-RSD currency was requested but no\nexchange rate was available, so amounts fall back to RSD.",
                    "type": "string",
                    "example": "Requested EUR, but no exchange rate available; showing RSD"
                },
                "generated_at": {
                    "type": "string",
                    "example": "2025-12-06T18:30:00Z"
                },
                "period": {
                    "$ref": "#/definitions/dto.ReportPeriod"
                },
                "report_type": {
                    "type": "string",
                    "example": "spending_by_category"
                },
                "spending_by_category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CategorySpending"
                    }
                },
                "total_amount": {
                    "type": "number",
                    "example": 58975
                },
                "total_transactions": {
                    "type": "integer",
                    "example": 15
                },
                "transaction_type": {
                    "type": "string",
                    "example": "expense"
                }
            }
        },
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.TransactionCounts": {
            "type": "object",
            "properties": {
                "expense_transactions": {
                    "type": "integer",
                    "example": 15
                },
                "income_transactions": {
                    "type": "integer",
                    "example": 5
                },
                "total_transactions": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
...
        "dto.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid email format"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}